refactor: relocate BackgroundRefresh DTOs to Application contracts
Move RefreshTaskDto, RefreshTaskStatusDto, and RefreshTaskExecutionLogDto
from Anela.Heblo.API/Controllers/ to
Anela.Heblo.Application/Features/BackgroundJobs/Contracts/, restoring
compliance with the rule that the API project never owns DTOs
(docs/architecture/development_guidelines.md). Update
BackgroundRefreshController's using directive accordingly. No behavior,
wire contract, or generated TypeScript client shape changes.
